backend/requirements.txt
fastapi
uvicorn
pydantic[email]
motor
python-decouple
PyJWT
passlib
bcrypt==4.0.1
python-dotenv
python-dateutil